refactor: 평가 점수 double로 수정, 지원서별 평가 List로 반환

+++ src/main/java/com/piveguyz/empickbackend/employment/jobtests/evaluation/query/dto/JobTestEvaluationDetailDTO.java
@@ -14,8 +14,8 @@ import java.util.List;
 @Setter
 public class JobTestEvaluationDetailDTO {
     private String evaluationComment;    // 총 코멘트 (application_job_test.evaluator_comment)
-    private Integer evaluationScore;     // 최종 평가 점수 (application_job_test.evaluation_score)
-    private Integer gradingTotalScore;   // 채점 총점 (application_job_test.grading_total_score)
+    private Double evaluationScore;     // 최종 평가 점수 (application_job_test.evaluation_score)
+    private Double gradingTotalScore;   // 채점 총점 (application_job_test.grading_total_score)
     private String evaluatorName;        // 평가자 이름 (member.name)
 
     private List<EvaluationResultQueryDTO> evaluationResults; // 각 기준별 결과 목록
